feat: extract value evaluator utility
